backend/app/parsers.py
from fastapi import UploadFile
import pdfminer
import docx
import pandas as pd
from bs4 import BeautifulSoup

# Sample parser for PDF
async def parse_pdf(file: UploadFile) -> str:
    # Implement PDF to Markdown conversion using pdfminer or similar libraries
    return "# PDF Content Parsed as Markdown"

# Sample parser for DOCX
async def parse_docx(file: UploadFile) -> str:
    # Implement DOCX to Markdown conversion using python-docx
    return "# DOCX Content Parsed as Markdown"

# Sample parser for CSV
async def parse_csv(file: UploadFile) -> str:
    # Implement CSV to Markdown conversion using pandas
    return "# CSV Content Parsed as Markdown"

# Sample parser for HTML
async def parse_html(file: UploadFile) -> str:
    # Implement HTML to Markdown conversion using BeautifulSoup
    return "# HTML Content Parsed as Markdown"
