backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
beautifulsoup4==4.12.3
blinker==1.8.2
click==8.1.7
fastapi==0.115.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==2.1.5
numpy==2.1.1
pandas==2.2.3
pdfminer==20191125
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-multipart==0.0.12
pytz==2024.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.38.6
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.31.0
Werkzeug==3.0.4
